+++ src/webui/components/RegistryInfoContent/styles.js
@@ -1,7 +1,0 @@
-import styled from 'react-emotion';
-
-export const CommandContainer = styled.div`
-  && {
-    padding-top: 20px;
-  }
-`;