+++ src/webui/components/Tag/types.js
@@ -1,9 +1,0 @@
-/**
- * @prettier
- * @flow
- */
-import type { Node } from 'react';
-
-export interface IProps {
-  children: Node;
-}