+++ src/webui/utils/constants.js
@@ -1,9 +1,0 @@
-export const TEXT = {
-  CLIPBOARD_COPY: 'Copy to Clipboard',
-};
-
-export const NODE_MANAGER = {
-  npm: 'npm',
-  yarn: 'yarn',
-  pnpm: 'pnpm',
-};