+++ src/webui/utils/windows.js
@@ -1,3 +1,0 @@
-export function goToVerdaccioWebsite() {
-	window.open('https://www.verdaccio.org/', '_blank');
-}