+++ test/e2e/setup.js
@@ -1,4 +1,4 @@
-const chalk = require('chalk');
+const { green } = require('kleur');
 const puppeteer = require('puppeteer');
 const fs = require('fs');
 const mkdirp = require('mkdirp');
@@ -8,8 +8,8 @@ const path = require('path');
 const DIR = path.join(os.tmpdir(), 'jest_puppeteer_global_setup');
 
 module.exports = async function() {
-  console.log(chalk.green('Setup Puppeteer'));
-  const browser = await puppeteer.launch({headless: true, /* slowMo: 300 */ });
+  console.log(green('Setup Puppeteer'));
+  const browser = await puppeteer.launch({ headless: true, /* slowMo: 300 */ args: ['--no-sandbox'] });
   global.__BROWSER__ = browser;
   mkdirp.sync(DIR);
   fs.writeFileSync(path.join(DIR, 'wsEndpoint'), browser.wsEndpoint());