+++ test/functional/readme/readme.js
@@ -5,6 +5,7 @@ export default function (server, server2) {
   describe('should test readme', () => {
     const README_PKG1 = 'readme-test';
     const README_PKG2 = 'readme-test-no-readme';
+    const README_MESSAGE = 'this is a readme';
 
     beforeAll(async function() {
       await server.putPackage('readme-test', require('./pkg-readme.json'))
@@ -16,20 +17,21 @@ export default function (server, server2) {
     test('add pkg', () => {});
 
     describe('should check readme file', () => {
-      const matchReadme = (serverRef, pkgName = README_PKG1, readmeMessage = 'this is a readme') => {
+      const matchReadme = (serverRef, pkgName = README_PKG1, readmeMessage = README_MESSAGE) => {
         return serverRef.request({
           uri: `/-/verdaccio/package/readme/${pkgName}`
         }).status(HTTP_STATUS.OK).then(function(body) {
-          expect(body).toEqual(`<p>${readmeMessage}</p>\n`);
+
+          expect(body).toEqual(`<p>${readmeMessage}</p>`);
         });
       };
 
       test('should fetch server2 over uplink server1', () => {
-        return matchReadme(server, README_PKG1, 'this is a readme');
+        return matchReadme(server, README_PKG1, README_MESSAGE);
       });
 
       test('should fetch package on local server1', () => {
-        return matchReadme(server2, README_PKG1, 'this is a readme');
+        return matchReadme(server2, README_PKG1, README_MESSAGE);
       });
 
       test('should fetch not found readme server2 over uplink server1', () => {