+++ test/functional/teardown.js
@@ -1,4 +1,4 @@
 module.exports = async function() {
-  // console.log(chalk.green('Teardown Verdaccio Functional'));
-  // console.log(chalk.blue('Teardown:: all server were closed'));
+  // console.log(green('Teardown Verdaccio Functional'));
+  // console.log(blue('Teardown:: all server were closed'));
 };