+++ test/unit/api/utils.spec.js
@@ -14,7 +14,8 @@ import {
   getWebProtocol,
   getVersionFromTarball,
   sortByName,
-  formatAuthor
+  formatAuthor,
+  isHTTPProtocol,
 } from '../../../src/lib/utils';
 import { DIST_TAGS, DEFAULT_USER } from '../../../src/lib/constants';
 import Logger, { setup } from '../../../src/lib/logger';
@@ -332,6 +333,26 @@ describe('Utilities', () => {
 
       expect(url).toMatch('/-/static/logo.png');
     });
+
+    test('should check HTTP protocol correctly', () => {
+      expect(isHTTPProtocol('http://domain.com/-/static/logo.png')).toBeTruthy();
+      expect(isHTTPProtocol('https://www.domain.com/-/static/logo.png')).toBeTruthy();
+      expect(isHTTPProtocol('//domain.com/-/static/logo.png')).toBeTruthy();
+      expect(isHTTPProtocol('file:///home/user/logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('file:///F:/home/user/logo.png')).toBeFalsy();
+      // Note that uses ftp protocol in src was deprecated in modern browsers
+      expect(isHTTPProtocol('ftp://1.2.3.4/home/user/logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('./logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('.\\logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('../logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('..\\logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('../../static/logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('..\\..\\static\\logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('.logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('/static/logo.png')).toBeFalsy();
+      expect(isHTTPProtocol('F:\\static\\logo.png')).toBeFalsy();
+    });
   });
 
   describe('User utilities', () => {
@@ -353,7 +374,7 @@ describe('Utilities', () => {
     test('should parse makrdown text to html template', () => {
       const markdown = '# markdown';
       expect(parseReadme('testPackage', markdown)).toEqual(
-        '<h1 id="markdown">markdown</h1>\n'
+        '<h1 id="markdown">markdown</h1>'
       );
       expect(
         parseReadme('testPackage', String(readmeFile('markdown.md')))
@@ -366,13 +387,13 @@ describe('Utilities', () => {
       const randomTextMarkdown = 'simple text \n # markdown';
 
       expect(parseReadme('testPackage', randomText)).toEqual(
-        '<p>%%%%%**##==</p>\n'
+        '<p>%%%%%**##==</p>'
       );
       expect(parseReadme('testPackage', simpleText)).toEqual(
-        '<p>simple text</p>\n'
+        '<p>simple text</p>'
       );
       expect(parseReadme('testPackage', randomTextMarkdown)).toEqual(
-        '<p>simple text </p>\n<h1 id="markdown">markdown</h1>\n'
+        '<p>simple text </p>\n<h1 id="markdown">markdown</h1>'
       );
     });
 
@@ -380,7 +401,7 @@ describe('Utilities', () => {
       const noData = '';
       const spy = jest.spyOn(Logger.logger, 'error');
       expect(parseReadme('testPackage', noData)).toEqual(
-        '<p>ERROR: No README data found!</p>\n'
+        '<p>ERROR: No README data found!</p>'
       );
       expect(spy).toHaveBeenCalledWith(
         { packageName: 'testPackage' },